Add files via upload
Online experiment during which participants are first shown a menu to familiarize them with the food stimuli. Then a total of 40 food items is first rated on desirability (i.e., "How much would you like to eat this item?") during a rating phase. Participants then rate their happiness on a 9-point and their hunger on a 7-point Likert scale. During the following choice phase, 120 choice options are shown between two food items. In an attention manipulation, right before a food item is shown, a directional (a left- or right-pointing arrow) or neutral (a bidirectional arrow) is presented, akin to a Posner cueing task. Food items could be healthier, natural and unprocessed foods, or unhealthier ultra-processed foods. This is an online task making use of OpenSesameWeb, that could be for example hosted on a JATOS server. This task was used in the following paper:
Schomaker, J., Vriens, M., & Jarva, H. A. (2022). Healthy or not: influencing attention to bias food choices. Food Quality and Preference, 96, 104384.